move healthcheck to its own file
